feat(feelers): expose essential parse utilities
Mirror feelin - exposes the parse feature, too.

Otherwise tools that use feelers AND do source code analysis must
explicitly depend on upstream libraries - potentially causing drift.
