+++ packages/feelers/test/spec/index.spec.js
@@ -1,12 +1,22 @@
 import { expect } from 'chai';
 
-import { evaluate } from 'feelers';
+import { evaluate, parse, parseToSimpleTree } from 'feelers';
 
 
 describe('feelers', function() {
 
   it('should export <evaluate>', function() {
     expect(evaluate).to.exist;
+  });
+
+
+  it('should export <parse>', function() {
+    expect(parse).to.exist;
+  });
+
+
+  it('should export <parseToSimpleTree>', function() {
+    expect(parseToSimpleTree).to.exist;
   });
 
 });
